Remove Providers queries. (#24)
Simplifies bitswap quite a bit. The reason I think the providers query is a bad idea is because it depends on how the data is replicated. The current behaviour can be emulated by retrying failed queries with a different provider set.

For actyx all nodes replicate all data and are connected to all peers, so the providers are simply the set of all connected peers. In ipfs-embed we will stop publishing individual blocks in the dht and instead publish which aliases we use. This will enable actyx to use the dht for partial replication in the future.